requirements.txt

kfp==2.7.0
google-cloud-aiplatform==1.43.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
os
joblib==1.3.2
google-cloud-storage==2.14.0
google-cloud-bigquery==3.19.0
pyarrow==14.0.2